Refine module docs and clean compose
