+++ examples/module_template/requirements.txt
@@ -1,2 +1,1 @@
 git+https://github.com/nashspence/home-index.git
-debugpy
